requirements.txt
fastapi==0.110.2
mysql_connector_repackaged==0.3.1
